Adding button to toolbar to expand editor to full screen
Based on feedback we got at ACE 2019, people thought it would be a good idea to include a toolbar button that could make the method editor full screen, so they could have more room to edit their code. 

This is a first pass at that functionality that uses the standard browser fullscreen functionality to expand the method editor. We'll use this for a bit and hopefully get feedback on how it feels, but we may need to adjust how much of the editor is visible or what element of the window we start on (e.g. we may want to keep in the Item window toolbar in fullscreen mode).

This code also shouldn't be Monaco-editor specific, so we can probably pull out these changes directly and make it a community project in and of itself.